# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aniket Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  coverImage?: string | null;
  technologies: string[];
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  featured: boolean;
  order: number;
  status?: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  metrics?: string | null;
  /** @nullable */
  problemStatement?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  lessonsLearned?: string | null;
  createdAt?: string;
}

export interface Blog {
  id: number;
  slug: string;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  coverImage?: string | null;
  tags?: string[];
  published: boolean;
  /** @nullable */
  readingTime?: number | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  company?: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactResult {
  success: boolean;
  message: string;
}

export interface Resume {
  id: number;
  fileName: string;
  version: string;
  downloadUrl: string;
  createdAt: string;
}

export interface GithubProfile {
  username: string;
  /** @nullable */
  bio: string | null;
  publicRepos: number;
  followers: number;
  following: number;
  avatarUrl: string;
  htmlUrl?: string;
  totalStars?: number;
  totalCommits?: number;
  topLanguages?: string[];
}

export interface GithubRepo {
  name: string;
  /** @nullable */
  description: string | null;
  url: string;
  stars: number;
  forks?: number;
  /** @nullable */
  language: string | null;
  topics?: string[];
}

export interface SearchResults {
  projects: Project[];
  blogs: Blog[];
  skills: string[];
}

export interface NewsletterInput {
  email: string;
  name?: string;
}

export interface NewsletterResult {
  success: boolean;
  message: string;
}

export type AnalyticsEventInputMetadata = { [key: string]: unknown };

export interface AnalyticsEventInput {
  event: string;
  page: string;
  metadata?: AnalyticsEventInputMetadata;
}

export interface AnalyticsEventResult {
  success: boolean;
}

export type GlobalSearchParams = {
q: string;
};