# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aniket Portfolio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
...
  /blogs/{slug}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        "404":
          description: Not found
...
  /analytics/event:
    post:
      operationId: trackEvent
      tags: [analytics]
      summary: Track analytics event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventInput"
      responses:
        "200":
          description: Event tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsEventResult"
components:
  schemas:
...
    Project:
      type: object
      required: [id, slug, title, description, technologies, featured, order]
      properties:
        id:
          type: number
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        technologies:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: number
        status:
          type: string
        category:
          type: ["string", "null"]
        metrics:
          type: ["string", "null"]
        problemStatement:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        lessonsLearned:
          type: ["string", "null"]
        createdAt:
          type: string
    Blog:
      type: object
      required: [id, slug, title, summary, published, createdAt]
      properties:
        id:
          type: number
        slug:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: ["string", "null"]
        coverImage:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        readingTime:
          type: ["number", "null"]
        createdAt:
          type: string
...
    AnalyticsEventInput:
      type: object
      required: [event, page]
      properties:
        event:
          type: string
        page:
          type: string
        metadata:
          type: object
          additionalProperties: true
    AnalyticsEventResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean