# lib/api-zod/src/generated/types/analyticsEventInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aniket Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsEventInputMetadata } from './analyticsEventInputMetadata';

export interface AnalyticsEventInput {
  event: string;
  page: string;
  metadata?: AnalyticsEventInputMetadata;
}