# lib/api-zod/src/generated/types/blog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aniket Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Blog {
  id: number;
  slug: string;
  title: string;
  summary: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  coverImage?: string | null;
  tags?: string[];
  published: boolean;
  /** @nullable */
  readingTime?: number | null;
  createdAt: string;
}