# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aniket Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  coverImage?: string | null;
  technologies: string[];
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  featured: boolean;
  order: number;
  status?: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  metrics?: string | null;
  /** @nullable */
  problemStatement?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  lessonsLearned?: string | null;
  createdAt?: string;
}